Add possible solution for Emacs border issue
There has been a black border on the right hand side of my Emacs window
when I have launched it. Or at least I thought it was black but it has
sometimes showed something that looked like text there as well making me
realize that this was probably not a setting.

I think this patch might solve the issue but it requires further testing.