Switch from pywal to a customized theme
pywal is very nice to automatically generate colorschemes based on a an
image. However I find myself easily getting caught in the not happy with
what I have situation. The possibility of finding the perfect picture
that makes everything come together.

So instead I will go down the route with customized colors and a low key
background that looks clean and cool and that I don't need to bother
customizing. Getting colors to fit between different programs is a
hassle so a dark background makes it easier on the eyes.