Move toward a Emacs daemon centric setup
This is a work in progress but my direction at the moment is to make
bspwm more and more as a thin layer which exists on top of an Emacs
daemon setup.

Empty desktops will provide me with quick hot keys to launch Emacs
frames that contains useful setups such as, todo(agenda view), (e)shell,
file commander (ranger), chat client (erc) etc.