Migrate from bspwm to stumpwm
StumpWM, dubbed the Emacs of WM's is built in Common Lisp. These days I
spend most of my time inside Emacs and are not really using all the
bells and whistles that bspwm brings to the table. So I rather have a
more hackable and WM built in lisp that integrates nicely with Emacs.